+++ src/main/java/com/example/Crud/API/Spring/Boot/controller/MedicoController.java
@@ -2,8 +2,11 @@ package com.example.Crud.API.Spring.Boot.controller;
 
 
 import com.example.Crud.API.Spring.Boot.model.Medico;
+import com.example.Crud.API.Spring.Boot.model.Paciente;
 import com.example.Crud.API.Spring.Boot.services.MedicoService;
+import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
@@ -32,21 +35,28 @@ public class MedicoController {
     }
 
     @PutMapping(path = "/{id}")
-    public Medico updateMedico( @RequestBody Medico request, @PathVariable("id") Long id) {
-        return this.medicoService.updateMedico(request, id);
+    public ResponseEntity updateMedico(@PathVariable("id") Long id, @RequestBody Medico request) {
+        Optional<Medico> existingMedico = this.medicoService.getMedicoById(id);
+        if (existingMedico.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        Medico updateMedico = this.medicoService.updateMedico(request, id);
+        return ResponseEntity.ok(updateMedico);
     }
 
     @DeleteMapping(path = "/{id}")
     public String deleteMedico(@PathVariable("id") Long id) {
-        boolean ok = this.medicoService.deleteMedico(id);
-        if (ok){
-            return "Medico con id " + id + " eliminado";
-        }else{
-            return "Error al eliminar el Medico";
+        if (!medicoService.getMedicoById(id).isPresent()) {
+            return "Error al eliminar el medico con id: " + id;
+        } else {
+            boolean ok = this.medicoService.deleteMedico(id);
+            if (ok) {
+                return "Medico con id " + id + " eliminado";
+            } else {
+                return "Error al eliminar el Medico";
+            }
         }
     }
-
-
 
 
 }
